feat: validate-tests: add option to set ID as test name
The new mrgvalidate input links tests between the specs and tests by
test ID, but packages that used the old mrgvalidate won't have test
IDs in the issues or test names.

For legacy support, let the validate_tests() caller specify that the
test ID column should be populated from the test names.  On the spec
side, test ID (e.g., from GitHub issues) also needs to be the name.

And mrgvalidate requires the ability to handle this (e.g.,
auto-assigning IDs) so that the output doesn't have ID and NAME info
repeated.

Note: I'm piggybacking on an existing test.  A dedicated one would be
better, but these validate_tests() tests are pretty slow (a bit under
4 minutes for me locally).