feat: parse-test-output: strip trailing colon when rolling up IDs
A good number of tests in bbr have tests that share a common ID and
use the pattern

  {test name}: {case} [ID]

In this case, the rolled up ID becomes "{test name}:".  The trailing
colon is unnecessary and potentially confusing, so teach leading_lcs()
to strip it along with the other trailing characters it already
strips.

Re: https://github.com/metrumresearchgroup/bbr/pull/430#discussion_r745814473