add function to run tests and write results
The r-dev-ci-mrgval image and various projects have scripts that do
roughly the same steps: run the tests, process the results with
parse_testthat_list_reporter(), and then write output for mrgvalidate
to consume.  This same logic will be needed for the upcoming "test
results to git" functionality, so add a function that implements these
steps and export it to hopefully reduce the duplication of this logic
across different workflows.

Note that there is also functionality for running tests in
validate-tests.R's internal run_tests().  Don't try to use that here
because, for the purpose of this new function, run_tests() seems to do
a bit more than needed (e.g., it invokes test_dir in a subprocess).
It might be possible to simplify run_tests(), but it's not worth the
effort given it's only used in legacy functionality that's slated to
be removed.

Also note that, like the r-dev-ci-mrgval image, the new function
defines the JSON data directly rather than using get_sys_info().  This
avoids the more indirect approach of setting environment variables
just so that get_sys_info() can retrieve them and allows using more
consistent casing in the field names.