add functions for storing test results in Git
The r-dev-ci-mrgval image runs a package's tests in Drone, generates
the validation docs, and uploads them to S3.  Importantly the image
bundles the reference docxs.  The downside of this approach is that it
only works for R packages and requires that the full test suite can be
executed in Drone, which is not the case for some R packages (e.g.,
some bbr tests are skipped on Drone because they require NONMEM).

To lift this constraint, the plan is to extend r-dev-ci-mrgval to
consume test results out of band.  In particular, the results will
stored and retrieved via a dedicated Git reference, where the
top-level trees are named by the tree IDs for the _tested_ commits.
Drone can look up the results for its current commit by its tree ID.

This commit implements the mrgval-prep side of that.  It exposes two
user-facing functions:

 * a function that runs the current repo's tests and pushes the
   results

 * a function that consumes a directory of results and pushes them

The former is meant for most R packages, while the latter makes it
possible to work with results from more complicated R packages
(perhaps mrgsolve fits in this category) or with results from other
languages.